requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-whois==0.8.0
reportlab==4.0.4
phonenumbers==8.13.19
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
dnspython<=2.0.0
ipwhois==1.2.0
werkzeug==2.3.7